requirements.txt
fastapi
websockets
google-generativeai
uvicorn
pillow  # For handling images if needed
pymongo